Ui: add subtils styles for end date
End date input is a specific case: if it's setted to the same day, you
probably don't really want this extra info. The only thing that care
is the starting and ending times. But if the event ends on another day,
or if your event is a all-day long one, then you should probably want
this information.

This commit sets it in the popover view: the end date really appear only
if it's matter.