Ui: improve the way captions are set in event popover
Event popover has a little bit complicated captions, because it need to
take care of many parameters: is the end date the same as the start date
and is the event is timed or all-day long. All of them can be combinated
which can results of many cases. This commit improve the way we handle
captions and refresh them.