Keep the event model unchanged until the user hits save. #465
Add an ugly but pretty safe context property, shared between all the event popover screens.

Also, there is an issue when there are one or more attendees to an event, the modal confirm dialog causes the model to be empty when the POST request is finally made. I was not able to identify the cause of this issue in a reasonnable time, so I implemented a dirty fix with temporary clone of the model attributes (see client/apps/models/scheduleitem.coffee.