fix behaviour issues on event popover confirm screen
Those issues are due to previous handling of "click out" events. Now the envent popover needs to be able to confirm closing every time a closing is requested. Also, this commit prepare the future handling of a callback called after the popover closing is actually done.

So from now, every time the popover close method is called, a verification of the formModel state is made to determine if the confirm screen has to be shown or not. Old dispayConfirmIfNeeded has been replaced by a confirmClose method, called conditionnaly from the close method.