Prevent automatic share
Before this PR the behavior was to always start the "share routine" on
the server side, even if an event is not shared.
With those changes the "share routine" resembles that of the mail: the
client side of the calendar application adds a parameter `share` to the
url to inform on whether or not the event is shared. The routine is then
launched only when needed.