No lookup error in DS for unexisting sharing
Before this commit when we wanted to get the sharing document associated
with a shared event we attempted a fetch of the document which id
matches the "shareID". The problem was that if we are the recipient of
the shared event no document with this id exists hence resulting in the
data-system returning an error and the client browser displaying a 404
in the console.

To avoid this behavior a new route was created on the server side of the
application to call the route "data/exist/:id" of the data-system.

Having that we now start by asking if a document with that id exists and
if that is the case we fetch it and if not we fetch the appropriate
sharing document. We no longer produce errors.